Refactor: Organize classes into categories and update module-info.java
- Moved classes into new packages: levels, displays, planes, and projectiles for better organization.
- Updated module-info.java to reflect new package structure with appropriate `opens` directives.
