+++ src/main/java/module-info.java
@@ -7,4 +7,7 @@ module com.example.demo {
     opens com.example.demo to javafx.fxml;
     exports com.example.demo.controller;
     opens com.example.demo.levels to javafx.fxml;
+    opens com.example.demo.displays to javafx.fxml;
+    opens com.example.demo.planes to javafx.fxml;
+    opens com.example.demo.projectiles to javafx.fxml;
 }
